mouse rotation hook
